Rename to version picker.
